requirements.txt
django
markdown
djangorestframework
django-filter
paho-mqtt
django-background-tasks